chunk-srclist.c: spell out the optimal strategy
This is the strategy described in my protva2018 report,
except that it limits the chunks to 4 records and thus uses
2 cut attempts.  Here the limit is 8 records with 6 cut attempts.

2.72 records per chunk
8.93% missing records
18.61% missing chunks

2.72 is actually exp(1), so it can't be further improved by increasing
the number of records in a chunk.  The math problem is worked out in
"How long until a random sequence decreases?" http://home.wlu.edu/~siehlerj/monotone.pdf
It boils down a formula similar to the second remarkable limit.