shuffle.c: initial revision, headers in random order
This implements a naive Fisher-Yates shuffle (there are small biases,
and not all permutations are equally likely - in fact, some permutations
are impossible).  This should be enough to estimate the effect on
compression, though.

$ set /var/lib/apt/lists/*list.*
$ echo $2
/var/lib/apt/lists/ftp.altlinux.org_pub_distributions_ALTLinux_Sisyphus_noarch_base_srclist.classic
$ zstd <$2 |wc -c
2742227
$ gcc -Wall -g -O2 shuffle.c -lzpkglist && ./a.out <$2 |zstd |wc -c
2955352

Note that these are source records, there are no subpackages which
would obviously improve compression a lot when grouped together.
The effect is entirely due to ordering of unrelated records by name.
Thus the name is a very strong predictor of local similarity.
Some time ago I actually tried to develop an even better predictor
which analyzes the data directly (with a variant of shingling).
I wasn't able to produce a stronger predictor, though (which would
deliver better compression).  The simple predictor which uses names
is already very strong.

The above demonstrates the improvement of about 7%.  The result turns
out too optimistic, because of fast compression which is not as good at
finding distant matches.

$ xz -d <data/12-noarch-srclist.classic.xz |
  ./compress-srclist -19 -D tmp/noarch.src.dict 2>/dev/null |wc -c
2335998
$ xz -d <data/12-noarch-srclist.classic.xz |./a.out |
  ./compress-srclist -19 -D tmp/noarch.src.dict 2>/dev/null |wc -c
2396724

Now the improvement is only about 2.5%.  This is still non-negligible,
and actually a lot, from the perspective of high compression.  However,
the total improvement which comes from grouping records was estimated
to be about 8-10% (in my protva2018 paper).  This means that local
similarity explains *less than half* of this total improvement,
while simply grouping anything with anything else explains more than
half of the total improvement.