test-compress-pkglist: show the number of chunks
Other things being equal, we want fewer chunks.  Furthermore, we can
introduce trade-offs.  For example, we can reluctantly permit 1%
increase in the number of missing chunks, provided that the total number
of chunks goes down by 10%.  In other words, the slope is 10, we get
a lot of something else in return, and that something else is not
unimportant.

noarch 55063884->6817086 bytes 18865->4939 chunks  missing: 1291505 19.43% bytes 750 15.19% chunks
x86_64 97619611->20693116 bytes 21030->4756 chunks  missing: 5234013 25.52% bytes 886 18.64% chunks