chunker.c: initial revision
This is a new chunker, it looks ahead for kindred records and tries
to put them in separate chunks.  When there is a single intermittent
record, AABCC..., it does AAB|CC...  When there are a few intermittent
records, it places them in a separate chunk: AABCDD... -> AA|BC|DD...
However, when there is a long enough run of stray records, it applies
the usual q[1] > q[2] strategy.  Thus it switches smoothly between the
two strategies!  The bulk of the code / a first non-working version :-)
was written in September 2018.

-noarch 55063884->6821535 bytes 18865->4997 chunks  missing: 1282743 19.28% bytes 758 15.18% chunks
+noarch 55063884->6888087 bytes 18865->6506 chunks  missing: 1088216 16.20% bytes 828 12.74% chunks

-x86_64 97619611->20694292 bytes 21030->4782 chunks  missing: 5225538 25.48% bytes 892 18.65% chunks
+x86_64 97619611->20758654 bytes 21030->6390 chunks  missing: 4456072 21.66% bytes 943 14.77% chunks

So the strategy increases the size by 1% and 0.3% (for pkglist.noarch
and x86_64 respectively), but the number of missing chunks goes down
by 17%!  That's a very nice slope.  This also increases the number of
chunks by 30%, but that's a secondary parameter.