compress-srclist.c: split 4 into 2+2
-x86_64 10405231->2327873 bytes 9377->3598 chunks  missing: 480421 22.28% bytes 734 20.41% chunks
+x86_64 10405231->2342961 bytes 9377->4043 chunks  missing: 446408 20.56% bytes 769 19.03% chunks

-noarch 11866651->2467238 bytes 11017->4190 chunks  missing: 429251 18.65% bytes 693 16.55% chunks
+noarch 11866651->2485890 bytes 11017->4728 chunks  missing: 396769 17.10% bytes 728 15.41% chunks

The total size increases by 0.7%, but the size of missing chunks
decreases by 7%.  This is a good slope.  However, the number of chunks
increases by 12%.  We can carry on in a similar vein by simply packaging
each record separately: the total size will increase by 8-9%, the size
of missing chunks will go down by 1.75x, and the total number of chunks
will be 2.32x.  This would be too radical a change, though.