compress-srclist.c: simplify the 2-3 strategy, issue 2 early
The next logical step is just to issue 2 early, when 3 doesn't work.
The strategy, in fact, produces similar results.  The main difference
is that it issues 2+3 while the previous strategy issues 3+2.  It also
issues 2+2+2 instead of 3+3, but the stretches of length 6 are already
rare.  I like the strategy because it makes the decision locally, with
less look-ahead, and therefore can be easier integrated with the
strategy in chunker.c.

-x86_64 10405231->2342961 bytes 9377->4043 chunks  missing: 446408 20.56% bytes 769 19.03% chunks
+x86_64 10405231->2344098 bytes 9377->4073 chunks  missing: 445438 20.51% bytes 772 18.96% chunks
-noarch 11866651->2485890 bytes 11017->4728 chunks  missing: 396769 17.10% bytes 728 15.41% chunks
+noarch 11866651->2486597 bytes 11017->4751 chunks  missing: 395387 17.03% bytes 731 15.39% chunks