chunker.c: stick stray records to the shorter of the two runs
Consider the case of AAAABCC(?!C), which gets chunked into AAAAB|CC.
It is arguably better to chunk it into AAAA|BCC.  However, this can
only be done if CC is followed by DD or DEF.  Otherwise, if it's DEE,
then D has the preemptive right to stick to CC.  Thus the decision is
non-local, it depends on what happens next at a larger distance.

The result is about the same, though.  (And the number of chunks
is obviously the same.  What I was hoping for is that *smaller* chunks
would be missing.  This does not happen in practice, though.  Instead,
the number of missing chunks increases by 1 or 2, probably because
the non-local decision makes the overall strategy less stable.
The difference is close to the noise floor anyway.  So I probably
won't keep this change.  But at least I tired.)

-noarch 55063884->6914301 bytes 18865->6856 chunks  missing: 1047767 15.53% bytes 844 12.32% chunks
+noarch 55063884->6913964 bytes 18865->6856 chunks  missing: 1047243 15.53% bytes 846 12.35% chunks

-x86_64 97619611->20774898 bytes 21030->6592 chunks  missing: 4400437 21.37% bytes 958 14.54% chunks
+x86_64 97619611->20777192 bytes 21030->6592 chunks  missing: 4395521 21.35% bytes 959 14.56% chunks