chunker.c: separate stray records from long 5+ runs
Without subpackages, the chunker cuts the stream into 2 and 3 records
per chunk, which means that any single record can invalidate 2 or 3
neighboring records.  This does not hold with long runs, such as AAAAAB.
If B is changed, it will invalidate 5 A records, which is a bit too much.
One the other hand, the case of AAB is probably okay, it's not worse
than ABC.

So this is a conservative change which takes care of the A5+ cases.
I'm not sure whether the same thing should be done with A3 and A4.
(Statistically the cases A3 and A4 are similar, and they occur quite
a bit more often.  Splitting them can produce too many stray records.)

$ set /var/lib/apt/lists/*list.*
$ print -l $1 $4
/var/lib/apt/lists/ftp.altlinux.org_pub_distributions_ALTLinux_Sisyphus_noarch_base_pkglist.classic
/var/lib/apt/lists/ftp.altlinux.org_pub_distributions_ALTLinux_Sisyphus_x86%5f64_base_pkglist.classic
$ pkglist-query '%{SourceRPM}\n' $1 $4 |uniq -cd |awk '{print$1}' |sort |uniq -c |sort -n |tail
     29 11
     39 10
     52 9
     63 8
     77 7
    211 6
    300 5
   1190 4
   1205 3
   4514 2

-noarch 55063884->6914301 bytes 18865->6856 chunks  missing: 1047767 15.53% bytes 844 12.32% chunks
+noarch 55063884->6908300 bytes 18865->6912 chunks  missing: 1033535 15.34% bytes 844 12.22% chunks

-x86_64 97619611->20774898 bytes 21030->6592 chunks  missing: 4400437 21.37% bytes 958 14.54% chunks
+x86_64 97619611->20768318 bytes 21030->6653 chunks  missing: 4343457 21.10% bytes 957 14.39% chunks