test-pkglist-seeding: same seed vs different seeds
> The seed used during compression may be different from the seed
> used at dictionary creation time (resulting in different chunking).
> This program tests whether using the same seed improves compression
> (the answer is no).

The first column below is the compressed size with the same seed,
the second is with different seeds.  The criterion that I use is that,
if the two intervals overlap, there is no significant difference.
Otherwise, if the two intervals do not overlap, and furthermore the
first size is smaller, then the same seed should be used to improve
compression (this may ultimately affect the data format).  Now,
the criterion can make smile some statisticians, because they are
aware of much more elaborate criteria.  Nonetheless, I believe the
criterion is about right.

x86_64 20761101±15622 20761447±15849     (size increase 346)
noarch 6544650±8158 6545306±9521         (size increase 656)